LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-position-area-expected.txt


FAIL flip-inline, position-area:left top, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-inline, position-area:left bottom, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-inline, position-area:right bottom, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-inline, position-area:right top, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block, position-area:left top, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-block, position-area:left bottom, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block, position-area:right bottom, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-block, position-area:right top, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-block flip-inline, position-area:left top, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-block flip-inline, position-area:left bottom, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-block flip-inline, position-area:right bottom, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block flip-inline, position-area:right top, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-start, position-area:left top, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-start, position-area:left bottom, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-start, position-area:right bottom, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-start, position-area:right top, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-block flip-start, position-area:left top, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-block flip-start, position-area:left bottom, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block flip-start, position-area:right bottom, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-block flip-start, position-area:right top, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-inline flip-start, position-area:left top, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-inline flip-start, position-area:left bottom, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-inline flip-start, position-area:right bottom, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-inline flip-start, position-area:right top, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block flip-inline flip-start, position-area:left top, ltr horizontal-tb assert_equals: expected "right bottom" but got "none"
FAIL flip-block flip-inline flip-start, position-area:left bottom, ltr horizontal-tb assert_equals: expected "left bottom" but got "none"
FAIL flip-block flip-inline flip-start, position-area:right bottom, ltr horizontal-tb assert_equals: expected "left top" but got "none"
FAIL flip-block flip-inline flip-start, position-area:right top, ltr horizontal-tb assert_equals: expected "right top" but got "none"
FAIL flip-block flip-inline, position-area:span-left span-top, ltr horizontal-tb assert_equals: expected "span-right span-bottom" but got "none"
FAIL flip-inline, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-inline, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-inline, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-inline, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-block, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-block, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-block flip-inline, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-block flip-inline, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-block flip-inline, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block flip-inline, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-start, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-start, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-start, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-start, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-block flip-start, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-block flip-start, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block flip-start, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-block flip-start, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-inline flip-start, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-inline flip-start, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-inline flip-start, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-inline flip-start, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:x-start y-start, ltr horizontal-tb assert_equals: expected "x-end y-end" but got "none"
FAIL flip-block flip-inline flip-start, position-area:x-start y-end, ltr horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-block flip-inline flip-start, position-area:x-end y-end, ltr horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:x-end y-start, ltr horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-block flip-inline, position-area:span-x-start span-y-start, ltr horizontal-tb assert_equals: expected "span-x-end span-y-end" but got "none"
FAIL flip-block flip-inline, position-area:x-self-start y-self-start, ltr horizontal-tb assert_equals: expected "x-self-end y-self-end" but got "none"
FAIL flip-block flip-inline, position-area:span-x-self-start span-y-self-start, ltr horizontal-tb assert_equals: expected "span-x-self-end span-y-self-end" but got "none"
FAIL flip-inline, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-inline, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-inline, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-inline, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-block, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-block, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-block flip-inline, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-block flip-inline, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-block flip-inline, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block flip-inline, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-start, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-start, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-start, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-start, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-block flip-start, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-block flip-start, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block flip-start, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-block flip-start, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-inline flip-start, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-inline flip-start, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-inline flip-start, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-inline flip-start, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:block-start inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-end" but got "none"
FAIL flip-block flip-inline flip-start, position-area:block-end inline-start, ltr horizontal-tb assert_equals: expected "block-end inline-start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:block-end inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:block-start inline-end, ltr horizontal-tb assert_equals: expected "block-start inline-end" but got "none"
FAIL flip-block flip-inline, position-area:span-block-start span-inline-start, ltr horizontal-tb assert_equals: expected "span-block-end span-inline-end" but got "none"
FAIL flip-block flip-inline, position-area:self-block-start self-inline-start, ltr horizontal-tb assert_equals: expected "self-block-end self-inline-end" but got "none"
FAIL flip-block flip-inline, position-area:span-self-block-start span-self-inline-start, ltr horizontal-tb assert_equals: expected "span-self-block-end span-self-inline-end" but got "none"
FAIL , position-area:start end, ltr horizontal-tb assert_equals: expected "start end" but got "none"
FAIL flip-block, position-area:start end, ltr horizontal-tb assert_equals: expected "end" but got "none"
FAIL flip-inline, position-area:start end, ltr horizontal-tb assert_equals: expected "start" but got "none"
FAIL flip-block flip-inline, position-area:start end, ltr horizontal-tb assert_equals: expected "end start" but got "none"
FAIL flip-start, position-area:start, ltr horizontal-tb assert_equals: expected "start" but got "none"
FAIL flip-start, position-area:end, ltr horizontal-tb assert_equals: expected "end" but got "none"
FAIL flip-start, position-area:start end, ltr horizontal-tb assert_equals: expected "end start" but got "none"
FAIL flip-block flip-start, position-area:start end, ltr horizontal-tb assert_equals: expected "end" but got "none"
FAIL flip-inline flip-start, position-area:start end, ltr horizontal-tb assert_equals: expected "start" but got "none"
FAIL flip-block flip-inline flip-start, position-area:start end, ltr horizontal-tb assert_equals: expected "start end" but got "none"
FAIL flip-block flip-inline, position-area:span-start span-end, ltr horizontal-tb assert_equals: expected "span-end span-start" but got "none"
FAIL flip-block flip-inline, position-area:self-start self-end, ltr horizontal-tb assert_equals: expected "self-end self-start" but got "none"
FAIL flip-block flip-inline, position-area:span-self-start span-self-end, ltr horizontal-tb assert_equals: expected "span-self-end span-self-start" but got "none"
FAIL flip-block, position-area:left center, ltr horizontal-tb assert_equals: expected "left center" but got "none"
FAIL flip-block, position-area:center top, ltr horizontal-tb assert_equals: expected "center bottom" but got "none"
FAIL flip-block, position-area:center, ltr horizontal-tb assert_equals: expected "center" but got "none"
FAIL flip-block, position-area:start center, ltr horizontal-tb assert_equals: expected "end center" but got "none"
FAIL flip-block, position-area:center start, ltr horizontal-tb assert_equals: expected "center start" but got "none"
FAIL flip-inline, position-area:center start, ltr horizontal-tb assert_equals: expected "center end" but got "none"
FAIL flip-start, position-area:center start, ltr horizontal-tb assert_equals: expected "start center" but got "none"
FAIL flip-block, position-area:left span-all, ltr horizontal-tb assert_equals: expected "left" but got "none"
FAIL flip-block, position-area:span-all top, ltr horizontal-tb assert_equals: expected "bottom" but got "none"
FAIL flip-block, position-area:span-all, ltr horizontal-tb assert_equals: expected "span-all" but got "none"
FAIL flip-block, position-area:start span-all, ltr horizontal-tb assert_equals: expected "end span-all" but got "none"
FAIL flip-block, position-area:span-all start, ltr horizontal-tb assert_equals: expected "span-all start" but got "none"
FAIL flip-inline, position-area:span-all start, ltr horizontal-tb assert_equals: expected "span-all end" but got "none"
FAIL flip-start, position-area:span-all start, ltr horizontal-tb assert_equals: expected "start span-all" but got "none"
FAIL flip-block, position-area:left span-top, ltr horizontal-tb assert_equals: expected "left span-bottom" but got "none"
FAIL flip-inline, position-area:left span-top, ltr horizontal-tb assert_equals: expected "right span-top" but got "none"
FAIL flip-start, position-area:span-block-start inline-end, ltr horizontal-tb assert_equals: expected "block-end span-inline-start" but got "none"
FAIL flip-block, position-area:left top, ltr vertical-rl assert_equals: expected "right top" but got "none"
FAIL , position-area:x-start y-start, rtl horizontal-tb assert_equals: expected "x-start y-start" but got "none"
FAIL flip-block, position-area:x-start y-start, rtl horizontal-tb assert_equals: expected "x-start y-end" but got "none"
FAIL flip-inline, position-area:x-start y-start, rtl horizontal-tb assert_equals: expected "x-end y-start" but got "none"
FAIL flip-block, position-area:x-end y-start, ltr vertical-rl assert_equals: expected "x-start y-start" but got "none"
FAIL flip-inline, position-area:x-end y-start, ltr vertical-rl assert_equals: expected "x-end y-end" but got "none"
FAIL flip-inline, position-area:start end, rtl horizontal-tb assert_equals: expected "start" but got "none"
FAIL flip-inline, position-area:start end, ltr vertical-rl assert_equals: expected "start" but got "none"
FAIL flip-block, position-area:start end, rtl horizontal-tb assert_equals: expected "end" but got "none"
FAIL flip-block, position-area:start end, ltr vertical-rl assert_equals: expected "end" but got "none"
